+++ packages/create-graphback/src/init/communityTemplates.ts
@@ -1,6 +1,6 @@
 
 /**
- * Array containing extrernal templates for the Graphback
+ * Array containing external templates for the Graphback
  * See contribution guide for more information
  */
 export const externalTemplates = [
@@ -11,6 +11,15 @@ export const externalTemplates = [
       uri: 'https://github.com/aerogear/graphback',
       branch: 'templates-1.0.0',
       path: '/templates/ts-apollo-mongo-apache-kafka-backend'
+    }]
+  },
+  {
+    name: 'nestjs-apollo-mongodb-backend',
+    description: 'Apollo GraphQL server using NestJS framework and connecting to MongoDB database',
+    repos: [{
+      uri: 'https://github.com/namit-chandwani/nestjs-apollo-mongodb-backend',
+      branch: 'master',
+      path: '.'
     }]
   }
 ];